refactor: refactoring

+++ BACKEND/api-routes/books/delete.js
@@ -1,5 +1,6 @@
 const Books = require('../../database/models/Books');
 const responseSender = require('../../helpers/response-sender');
+const checkAccessType = require('../../helpers/check-access-type');
 
 const booksHandlerDelete = async (req, res) => {
   if (!req.query.id) {
@@ -7,9 +8,13 @@ const booksHandlerDelete = async (req, res) => {
   }
 
   try {
-    await Books.findByIdAndDelete({_id: req.query.id});
+    if (checkAccessType(req)) {
+      await Books.findByIdAndDelete({_id: req.query.id});
+      responseSender(res, 200, 'Book has been deleted!');
+    } else {
+      responseSender(res, 403, 'Forbidden!');
+    }
 
-    responseSender(res, 200, 'Book has been deleted!');
   } catch (err) {
     responseSender(err, 500, err.message);
   }
